[Fabric] Initial support to host React Native as a ContentIsland (#12227)
* Initial Islands support

* Hook up AutomationProvider when running as an island

* Input args from window message should be translated to scaled coords

* Change files

* revert packages.lock

* fix

+++ vnext/Microsoft.ReactNative/Fabric/Composition/CompositionHwndHost.h
@@ -38,7 +38,7 @@ struct CompositionHwndHost : CompositionHwndHostT<CompositionHwndHost> {
   void CreateDesktopWindowTarget(HWND window);
   void CreateCompositionRoot();
   void UpdateSize() noexcept;
-  double ScaleFactor() noexcept;
+  float ScaleFactor() noexcept;
 
   HWND m_hwnd;
   winrt::Microsoft::ReactNative::CompositionRootView m_compRootView{nullptr};
